Extract the mocks into their own submodule / subdirectory.
