+++ local-modules/doc-server/tests/test_BaseComplexMember.js
@@ -9,8 +9,7 @@ import { BaseComplexMember } from 'doc-server';
 
 import { Codec } from 'codec';
 import { FileAccess } from 'doc-server';
-
-import MockFile from './MockFile';
+import { MockFile } from 'doc-server/mocks';
 
 describe('doc-server/BaseComplexMember', () => {
   describe('constructor()', () => {
